+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Select all products/clients on depot change or all data realod if no prior selection exist (or couldn't find product/client)
 - Make saved searches dialog resizable
 - Show hidden files and folders in File Chooser
+- Copy element of list selection dialog into extra field on double click
 
 ### Add
 - Show a point in the Health Check icon to show if there is a warning/error in the Health Check